Refresh Token Structure Implemented

+++ src/controllers/CategoryController.ts
@@ -6,8 +6,6 @@ import categoryService from "../services/categoryService";
 import CustomError from "../lib/Error";
 import { HTTP_CODES } from "../config/enum";
 import { DEFAULT_LANG } from "../config";
-import { log } from "console";
-// eslint-disable-next-line @typescript-eslint/no-require-imports
 const i18n = new (require("../lib/i18n"))(DEFAULT_LANG);
 const logger = require('./logger');
 
@@ -16,7 +14,7 @@ const getCategories=async(req:Request,res:Response)=>{
 
     const categories = await categoryService.getAll();
 
-    if(!categories || categories.length === 0) throw new CustomError({code:HTTP_CODES.NOT_FOUND,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user.language, ["categories"])});
+    if(!categories || categories.length === 0) throw new CustomError({code:HTTP_CODES.NOT_FOUND,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user?.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user?.language, ["categories"])});
         
     logger.info(`Categories fetched: ${categories.length}`);
 
@@ -27,11 +25,11 @@ const getCategoryById=async(req:Request,res:Response)=>{
 
     const categoryId=req.body.id;
 
-    //if(!categoryId) throw new CustomError({code:HTTP_CODES.BAD_REQUEST,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user.language, ["id"])});
+    if(!categoryId) throw new CustomError({code:HTTP_CODES.BAD_REQUEST,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user?.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user?.language, ["id"])});
 
     const category=await categoryService.getById(categoryId);
 
-    if(!category) throw new CustomError({code:HTTP_CODES.NOT_FOUND,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user.language, ["category"])});
+    if(!category) throw new CustomError({code:HTTP_CODES.NOT_FOUND,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user?.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user?.language, ["category"])});
 
     logger.info(`Category Updated: ${categoryId}`);
 
@@ -44,7 +42,7 @@ const createCategory=async(req:Request,res:Response)=>{
     const body=req.body;
 
 
-    //if (!body.name) throw new CustomError({code:HTTP_CODES.BAD_REQUEST,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user.language, ["name"])});
+    if (!body.name) throw new CustomError({code:HTTP_CODES.BAD_REQUEST,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user?.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user?.language, ["name"])});
 
     const result=await categoryService.create(body);
 
@@ -60,13 +58,14 @@ const updateCategory=async(req:Request,res:Response)=>{
 
     const categoryId=req.body.id;
 
-    //if(!categoryId) throw new CustomError({code:HTTP_CODES.BAD_REQUEST,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user.language, ["id"])});
+    if(!categoryId) throw new CustomError({code:HTTP_CODES.BAD_REQUEST,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user?.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user?.language, ["id"])});
 
-    //if (!body.name) throw new CustomError({code:HTTP_CODES.BAD_REQUEST,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user.language, ["name"])});
+    if (!body.name) throw new CustomError({code:HTTP_CODES.BAD_REQUEST,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user?.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user?.language, ["name"])});
 
     if (body.name) updates.name = body.name;
     if (typeof body.is_active === "boolean") updates.is_active = body.is_active;
-    //if(req.user) updates.createdby=req.user.id
+    
+    if(req.user) updates.createdby=req.user.id
 
     const result =await categoryService.update(categoryId,updates)
 
@@ -79,7 +78,7 @@ const deleteCategory=async(req:Request,res:Response)=>{
 
     const categoryId=req.body.id;
 
-    //if(!categoryId) throw new CustomError({code:HTTP_CODES.BAD_REQUEST,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user.language, ["id"])});
+    if(!categoryId) throw new CustomError({code:HTTP_CODES.BAD_REQUEST,message: i18n.translate("COMMON.VALIDATION_ERROR_TITLE", req.user?.language),description: i18n.translate("COMMON.FIELD_MUST_BE_FILLED", req.user?.language, ["id"])});
 
 
     const result=await categoryService.delete(categoryId);
